Reduce alpha vector backup to a single matrix multiplication.

Current performance is quite bad relative to main branch, but I think the idea may still be worth pursuing.

# Before

### Tiger

```julia
BenchmarkTools.Trial: 558 samples with 1 evaluation.
 Range (min … max):  8.515 ms …  15.210 ms  ┊ GC (min … max): 0.00% … 41.33%
 Time  (median):     8.844 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   8.966 ms ± 833.336 μs  ┊ GC (mean ± σ):  1.29% ±  5.52%

    ▇█                                                         
  ▃▆██▇▄▃▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂▁▁▂▂ ▂
  8.52 ms         Histogram: frequency by time        14.7 ms <

 Memory estimate: 2.00 MiB, allocs estimate: 13031.
```

### Rocksample(5,5)

```julia
BenchmarkTools.Trial: 25 samples with 1 evaluation.
 Range (min … max):  201.661 ms … 209.820 ms  ┊ GC (min … max): 0.00% … 2.03%
 Time  (median):     205.455 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   205.419 ms ±   2.128 ms  ┊ GC (mean ± σ):  0.25% ± 0.69%

  ▁      ▁  ▁█▁▁▁    ▁ ▁  ▁  ▁▁   ███ ▁           █▁   ▁      ▁  
  █▁▁▁▁▁▁█▁▁█████▁▁▁▁█▁█▁▁█▁▁██▁▁▁███▁█▁▁▁▁▁▁▁▁▁▁▁██▁▁▁█▁▁▁▁▁▁█ ▁
  202 ms           Histogram: frequency by time          210 ms <

 Memory estimate: 14.99 MiB, allocs estimate: 69304.
```

### Rocksample(10,10)

```julia
BenchmarkTools.Trial: 17 samples with 1 evaluation.
 Range (min … max):  293.601 ms … 303.135 ms  ┊ GC (min … max): 0.00% … 1.79%
 Time  (median):     297.885 ms               ┊ GC (median):    0.71%
 Time  (mean ± σ):   298.023 ms ±   3.036 ms  ┊ GC (mean ± σ):  0.85% ± 0.90%

  █ █  █     █ ██    █     █ █ ██       █       █ █    ██     █  
  █▁█▁▁█▁▁▁▁▁█▁██▁▁▁▁█▁▁▁▁▁█▁█▁██▁▁▁▁▁▁▁█▁▁▁▁▁▁▁█▁█▁▁▁▁██▁▁▁▁▁█ ▁
  294 ms           Histogram: frequency by time          303 ms <

 Memory estimate: 64.93 MiB, allocs estimate: 82768.
```

# After

### Tiger

```julia
BenchmarkTools.Trial: 556 samples with 1 evaluation.
 Range (min … max):  8.625 ms …  16.160 ms  ┊ GC (min … max): 0.00% … 43.59%
 Time  (median):     8.874 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   9.002 ms ± 880.114 μs  ┊ GC (mean ± σ):  1.21% ±  5.39%

   █▅▁                                                         
  ▇███▅▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂▁▁▁▂ ▂
  8.62 ms         Histogram: frequency by time        15.7 ms <

 Memory estimate: 2.00 MiB, allocs estimate: 13043.
```

### Rocksample(5,5)

```julia
BenchmarkTools.Trial: 25 samples with 1 evaluation.
 Range (min … max):  200.425 ms … 208.799 ms  ┊ GC (min … max): 0.00% … 2.23%
 Time  (median):     203.231 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   203.680 ms ±   2.312 ms  ┊ GC (mean ± σ):  0.34% ± 0.78%

                  █            ▃   █                          ▃  
  ▇▁▇▇▇▁▁▁▇▁▇▁▇▇▁▁█▇▁▁▇▇▁▁▁▇▇▁▁█▁▁▁█▁▁▁▁▁▁▁▁▁▇▁▁▁▁▁▇▁▁▁▁▁▁▁▁▁▁█ ▁
  200 ms           Histogram: frequency by time          209 ms <

 Memory estimate: 22.43 MiB, allocs estimate: 69343.
```

### Rocksample(10,10)

```julia
BenchmarkTools.Trial: 17 samples with 1 evaluation.
 Range (min … max):  298.811 ms … 305.206 ms  ┊ GC (min … max): 0.71% … 1.33%
 Time  (median):     302.556 ms               ┊ GC (median):    1.37%
 Time  (mean ± σ):   302.161 ms ±   1.666 ms  ┊ GC (mean ± σ):  1.45% ± 0.33%

                    ▃                █        ▃                  
  ▇▁▁▁▁▁▁▇▁▁▁▁▁▁▁▁▁▁█▁▁▇▁▇▁▁▁▇▁▁▁▁▁▁▁█▇▇▁▇▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▇▁▁▁▁▇ ▁
  299 ms           Histogram: frequency by time          305 ms <

 Memory estimate: 182.57 MiB, allocs estimate: 82816.
```