@zsunberg 2023-02-28
Seems like there are a lot of allocs. Can you do it more in-place somehow?