@WhiffleFish 2023-02-28
A vast majority of allocations occur when generating new beliefs and storing these beliefs in the tree. While this could be improved by either preallocating beliefs at the very beginning or using something like a bump allocator around the belief updater, I'm not entirely sure it's worth the effort.

Currently, allocations have an extremely negligible effect on performance. In profiles, elementary non-allocating array operations like `getindex` and `setindex!` take up a vast majority of computation time.

As far as I can tell, the best ways of improving performance currently would be through #2, #3, and #7.

That being said, the current implementation is already quite fast, so these improvements aren't *absolutely* necessary.