@WhiffleFish 2023-03-01
`RockSample(10,10)` has 801 states, so there we're working with an 801 x 801 transition matrix.