@rejuvyesh 2023-03-01
If I understand the code correctly `T[sp, s]*Zᵀ[o, sp]` is the actual matrix multiplication? And you are still looping through things. I think unless we rewrite those loops somehow so that the result is from a few BLAS calls, there won't be any benefit of this.