@WhiffleFish 2023-03-01
Right, the line you're referencing is for initially populating the linear mapping matrix ($\beta$) before any planning begins. The actual alpha vector backup using this sparse matrix is done here:

https://github.com/JuliaPOMDP/NativeSARSOP.jl/blob/6baf19a90ad24d85abb0aa2fdb1c2e1efbd28b65/src/backup.jl#L14-L19
.